Update docs and use _resolve_filesystem_and_path
